+++ src/plugins/prettyexport/rsttable.h
@@ -7,10 +7,28 @@
  *
  */
 #include "prettyexport.h"
+#include <stdio.h>
 
-void calcSizes (PrettyHeadNode * head, PrettyIndexType indexType, ssize_t rowHeights[], ssize_t numRows, ssize_t colLengths[],
-		ssize_t numCols);
+struct _TableCell
+{
+	PrettyStyle style;
+	const char * value;
+};
+typedef struct _TableCell TableCell;
 
-ssize_t calcTableLength (ssize_t colLengths[], ssize_t numCols);
+void calcSizes (PrettyHeadNode * head, PrettyIndexType indexType, 
+        ssize_t numRows, ssize_t rowHeights[numRows], 
+        ssize_t numCols, ssize_t colLengths[numCols]);
 
-ssize_t calcTableHeight (ssize_t rowHeights[], ssize_t numRows);
+ssize_t calcTableLength (ssize_t numCols, ssize_t colLengths[numCols]);
+ssize_t calcTableHeight (ssize_t numRows, ssize_t rowHeights[numRows]);
+
+void callocTable (ssize_t tableLength, ssize_t tableHeight, TableCell * table[tableLength][tableHeight]);
+void freeTable (ssize_t tableLength, ssize_t tableHeight, TableCell * table[tableLength][tableHeight]);
+
+void printTableDebug(PrettyHeadNode * head, PrettyIndexType indexType, 
+        ssize_t tableLength, ssize_t tableHeight, TableCell * table[tableLength][tableHeight]);
+void printTableRst(FILE * fh, PrettyIndexNode * firstIndexNode, 
+        ssize_t tableLength, ssize_t tableHeight, TableCell * table[tableLength][tableHeight], 
+        ssize_t numCols, ssize_t colLengths[numCols], 
+        ssize_t numRows, ssize_t rowHeights[numRows]);
